requirements.txt
pandas
matplotlib
geopandas
streamlit
Pillow
